Move test running logic to /test/index.ts
